send UDP replies from the correct source address using IP_PKTINFO
For UDP protocols, when the server sends a reply datagram in response to
a request, it is important that the source address of the response
matches up with the destination address of the request. That is
necessary so that the datagram can be delivered to the right process on
the host which sent the request. It's also needed for any intermediate
routers or NAT devices which are doing connection tracking.

The traditional send() / sendto() calls on a UDP socket do not have any
way to specify the source address. On a host with a single IP address
this is not an issue. Even if there are multiple IP addresses, the
kernel has some heuristics to select the appropriate source address
based on the destination.

However on a multi-homed host (with overlapping routes) the kernel
heuristics are not sufficient. For example, if the host has two default
routes over two separate networks:

    # ip route
    default via 10.236.1.1 dev eth0 metric 1000
    default via 10.236.2.1 dev eth1 metric 1001

and a UDP request is received on eth1 on the 10.236.2.0/24 network, the
server must send its reply datagram out eth1 with the same source
address on 10.236.2.0/24. But the kernel will wrongly choose the address
on eth0 instead, because it has a lower metric.

We can specify a source address for a UDP datagram by using sendmsg()
with an IP_PKTINFO control message (IPV6_PKTINFO for IPv6).
Conveniently, if we set the corresponding sockopt() and then we will
also receive the destination address of an incoming datagram with the
same control message in recvmsg(), which we can pass through as is to
sendmsg().

This patch updates the built-in UDP servers to use recvmsg() / sendmsg()
and sets the relevant sockopt() to ensure the kernel delivers the
IP_PKTINFO control message when receiving.